+++ tests/CMakeLists.txt
@@ -5,7 +5,7 @@ include_directories(${PROJECT_SOURCE_DIR}/src
 
 list(APPEND tests_opentime test_opentime)
 foreach(test ${tests_opentime})
-	add_executable(${test} tests.h ${test}.cpp)
+	add_executable(${test} utils.h ${test}.cpp)
 	target_link_libraries(${test} opentime)
 	set_target_properties(${test} PROPERTIES FOLDER tests)
 	add_test(${test} ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/${test}${CMAKE_EXECUTABLE_SUFFIX})
@@ -13,7 +13,7 @@ endforeach()
 
 list(APPEND tests_opentimelineio test_clip)
 foreach(test ${tests_opentimelineio})
-	add_executable(${test} tests.h ${test}.cpp)
+	add_executable(${test} utils.h ${test}.cpp)
 	target_link_libraries(${test} opentimelineio)
 	set_target_properties(${test} PROPERTIES FOLDER tests)
 	add_test(${test} ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/${test}${CMAKE_EXECUTABLE_SUFFIX})
