+++ tests/test_clip.cpp
@@ -1,4 +1,4 @@
-#include "tests.h"
+#include "utils.h"
 
 #include <opentimelineio/clip.h>
 #include <opentimelineio/externalReference.h>
