+++ tests/test_opentime.cpp
@@ -1,4 +1,4 @@
-#include "tests.h"
+#include "utils.h"
 
 #include <opentime/rationalTime.h>
